// server/application/dto/goal-tracking.dto.ts
export type GoalTrackingDTO = {
  dailyQuestionGoal: number
  questionsCompletedToday: number
  weeklyStudyGoalMinutes: number
  weeklyStudyMinutesCompleted: number
  weeklyMockGoal: number
  weeklyMocksCompleted: number
  moduleCompletionGoal: number
  modulesCompleted: number
  completionPercentage: number
}